+++ tests/unit/util/test_configuration.py
@@ -1321,11 +1321,8 @@ metrics:
             """
 input:
   stdin:
-    type: file_input
-    logfile_path: tests/testdata/input_logdata/kafka_raw_event.jsonl
-    start: end
-    watch_file: true
-    interval: 1
+    type: jsonl_input
+    documents_path: tests/testdata/input_logdata/kafka_raw_event.jsonl
 """
         )